requirements.txt
psycopg2-binary>=2.9.9,<3.0
mcp>=1.0.0
python-dotenv>=1.0.0
python-telegram-bot>=20.0
openai>=1.90.0,<2.0.0
ollama>=0.4.7
pytest>=7.0
pytest-asyncio>=0.21.0
